machines: Use libosinfo APIs instead of calling its tools
Instead of calling libosinfo tools, let's rely on libosinfo APIs in
order to avoid doing a lot of parsing to match the detected OS from a
media or a tree.

In order to do so, the code in libvirt-common.js has been expanded to
also store the (unique) OS ID and, when detecting some media or tree,
just filter the list by it.

With this, two new python scripts have been added:
- autodetectOS.py, which does the media/tree detection and returns the
OS ID;
- getOSList.py, which gets exactly the same information (and in the same
format) got from osinfo-query, plus the OS ID.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>