machines: Return whether an OS is treeInstallable
This is a *hint* about whether the OS in question supports tree based
installations or not. Mind, though, that Tree installations may only
work with a specific variant of the system (and this is something that
has to be improved on virt-install's side).

The check done is basically:
- Check whether an OS has "treeinfo" information (case for RHEL, CentOS,
Fedora, and OpenSUSE) ... or .... check whether an OS has kernel &
initrd paths set up (case for Debian and Ubuntu).