machines: Allow getting unattended scripts info from medias
Let's add OsinfoMedia instances information to the getOSList script so
we can check whether a specific media supports or not unattended
installations and which profiles could be used by a specific media.

The reason medias are treated differently is that, for some distros,
some of the medias are not capable to deal with unattended
installations. Fedora is one of the main examples here, as its "live"
medias do not have anaconda running during the boot time, which means
there's nothing to read & interpret the kickstart file passed to the
kernel command line.

The logic to check whether a media supports or not unattended
installation is:
- Check whether the OsinfoOs supports unattended installation;
  - In case OsinfoOs does *not* support, an OsinfoMedia won't support
    either;
  - In case OsinfoOs supports unattended installation:
    - Check whether the OsinfoMedia supports the unattended
      installation;
      - In case the media has the "installer-script" set to false, the
        media does *not* support unattended installation;
      - Otherwise, it does;
        - In case it does, check which are the profiles supported;
    - In case there's no info about unattended scripts for a specific
      media, the ones supported by the OsinfoOs should be used.

This whole logic is a little bit tricky but, unfortunately, that's how
libosinfo evolved when having to deal with distros changes with
regarding to live / installable medias.